requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
starlette==0.37.2
tortoise_orm==0.20.1
pydantic==2.7.1
redis==5.0.4
python-multipart==0.0.9
python-dotenv==1.0.1
aerich==0.7.2
aiomysql==0.2.0
cryptography==42.0.8
passlib==1.7.4